backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0
psycopg>=3.1,<4.0
celery>=5.3,<6.0
redis>=5.0,<6.0
Pillow>=10.0,<11.0
django-storages>=1.14,<2.0
boto3>=1.34,<2.0
requests>=2.31,<3.0
dj-database-url>=2.1,<3.0
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
factory-boy>=3.3,<4.0
ruff>=0.5,<1.0
black>=24.0,<25.0
pre-commit>=3.7,<4.0
